requirements.txt
Django==4.2
psycopg2-binary==2.9.6
django-rest-framework
django-filter